chore(app, components, io): migrate over to HttpServer entirely
