+++ bin/agent-data-plane/src/internal/control_plane.rs
@@ -6,8 +6,7 @@ use arc_swap::ArcSwap;
 use datadog_agent_commons::ipc::tls::build_ipc_server_tls_config;
 use saluki_api::EndpointType;
 use saluki_app::{
-    accounting::ResourceTelemetryWorker, config::ConfigWorker, dynamic_api::DynamicAPIBuilder,
-    logging::LoggingOverrideController,
+    accounting::ResourceTelemetryWorker, api::APIBuilder, config::ConfigWorker, logging::LoggingOverrideController,
 };
 use saluki_core::accounting::ComponentRegistry;
 use saluki_core::{
@@ -56,12 +55,12 @@ pub async fn create_control_plane_supervisor(
     let api_listen_address = dp.api_listen_address()?;
     let secure_api_listen_address = dp.secure_api_listen_address()?;
 
-    supervisor.add_worker(DynamicAPIBuilder::new(EndpointType::Unprivileged, api_listen_address));
+    supervisor.add_worker(APIBuilder::new(EndpointType::Unprivileged, api_listen_address));
     let ipc_config = dp.ipc_auth_configuration();
     let tls_config = build_ipc_server_tls_config(ipc_config.ipc_cert_file_path()).await?;
 
     let mut privileged_api =
-        DynamicAPIBuilder::new(EndpointType::Privileged, secure_api_listen_address).with_tls_config(tls_config);
+        APIBuilder::new(EndpointType::Privileged, secure_api_listen_address).with_tls_config(tls_config);
 
     privileged_api = control_surfaces.register_control_surfaces(privileged_api);
