+++ lib/saluki-components/src/destinations/prometheus/mod.rs
@@ -1,8 +1,4 @@
 //! Prometheus destination.
-//!
-//! # Missing
-//!
-//! - Use `ShutdownCoordinator::shutdown_and_wait` so shutdown can be triggered and all HTTP connections can drain.
 
 use std::{
     convert::Infallible,
@@ -11,11 +7,11 @@ use std::{
 };
 
 use async_trait::async_trait;
+use axum::{extract::Request, Router};
 use ddsketch::DDSketch;
-use http::{Request, Response, StatusCode};
-use hyper::body::Incoming;
+use http::{Response, StatusCode};
 use prometheus_exposition::{MetricType, PrometheusRenderer};
-use saluki_common::{collections::FastIndexMap, iter::ReusableDeduplicator, sync::shutdown::ShutdownCoordinator};
+use saluki_common::{collections::FastIndexMap, iter::ReusableDeduplicator};
 use saluki_context::{tags::Tag, Context};
 use saluki_core::accounting::{MemoryBounds, MemoryBoundsBuilder};
 use saluki_core::components::{destinations::*, BuildContext};
@@ -23,12 +19,9 @@ use saluki_core::data_model::event::{
     metric::{Histogram, Metric, MetricValues},
     EventType,
 };
+use saluki_core::runtime;
 use saluki_error::GenericError;
-use saluki_io::net::{
-    listener::ConnectionOrientedListener,
-    server::http::{ErrorHandle, UnsupervisedHttpServer},
-    ListenAddress,
-};
+use saluki_io::net::{server::http::HttpServer, ListenAddress};
 use serde::Deserialize;
 use stringtheory::{
     interning::{FixedSizeInterner, Interner as _},
@@ -132,7 +125,7 @@ impl DestinationBuilder for PrometheusConfiguration {
 
     async fn build(&self, _context: BuildContext) -> Result<Box<dyn Destination + Send>, GenericError> {
         Ok(Box::new(Prometheus {
-            listener: ConnectionOrientedListener::from_listen_address(self.listen_addr.clone()).await?,
+            listen_addr: self.listen_addr.clone(),
             additional_routes: self.additional_routes.clone(),
             metrics: FastIndexMap::default(),
             payload: Arc::new(RwLock::new(String::new())),
@@ -161,7 +154,7 @@ impl MemoryBounds for PrometheusConfiguration {
 }
 
 struct Prometheus {
-    listener: ConnectionOrientedListener,
+    listen_addr: ListenAddress,
     additional_routes: Vec<PrometheusAdditionalRoute>,
     metrics: FastIndexMap<PrometheusContext, FastIndexMap<Context, PrometheusValue>>,
     payload: Arc<RwLock<String>>,
@@ -173,7 +166,7 @@ struct Prometheus {
 impl Destination for Prometheus {
     async fn run(mut self: Box<Self>, mut context: DestinationContext) -> Result<(), GenericError> {
         let Self {
-            listener,
+            listen_addr,
             additional_routes,
             mut metrics,
             payload,
@@ -183,8 +176,16 @@ impl Destination for Prometheus {
 
         let mut health = context.take_health_handle();
 
-        let (http_shutdown, mut http_error) =
-            spawn_prom_scrape_service(listener, Arc::clone(&payload), additional_routes);
+        // The scrape endpoint runs as a supervised worker of its own rather than as part of this component: the
+        // component's supervisor is what stops it, and drains its in-flight connections, once this component is done.
+        runtime::supervisable(build_scrape_server(
+            listen_addr,
+            Arc::clone(&payload),
+            additional_routes,
+        ))
+        .on_runtime(context.topology_context().global_thread_pool().clone())
+        .spawn();
+
         health.mark_ready();
 
         debug!("Prometheus destination started.");
@@ -236,18 +237,8 @@ impl Destination for Prometheus {
                     },
                     None => break,
                 },
-                error = &mut http_error => {
-                    if let Some(error) = error {
-                        debug!(%error, "HTTP server error.");
-                    }
-                    break;
-                },
             }
         }
-
-        // TODO: This should really use `ShutdownCoordinator::shutdown_and_wait` so we can trigger shutdown _and_ wait
-        // until all HTTP connections and the listener have finished.
-        http_shutdown.shutdown();
 
         debug!("Prometheus destination stopped.");
 
@@ -255,12 +246,15 @@ impl Destination for Prometheus {
     }
 }
 
-fn spawn_prom_scrape_service(
-    listener: ConnectionOrientedListener, payload: Arc<RwLock<String>>,
-    additional_routes: Vec<PrometheusAdditionalRoute>,
-) -> (ShutdownCoordinator, ErrorHandle) {
+/// Builds the server that answers scrape requests.
+///
+/// Every path is answered by the same service rather than being routed: the raw metrics paths take precedence over the
+/// additional routes, so a path claimed by both is answered with the raw payload rather than being a conflict.
+fn build_scrape_server(
+    listen_addr: ListenAddress, payload: Arc<RwLock<String>>, additional_routes: Vec<PrometheusAdditionalRoute>,
+) -> HttpServer {
     let additional_routes = Arc::new(additional_routes);
-    let service = service_fn(move |req: Request<Incoming>| {
+    let service = service_fn(move |req: Request| {
         let payload = Arc::clone(&payload);
         let additional_routes = Arc::clone(&additional_routes);
         async move {
@@ -268,8 +262,7 @@ fn spawn_prom_scrape_service(
         }
     });
 
-    let http_server = UnsupervisedHttpServer::from_listener(listener, service);
-    http_server.listen()
+    HttpServer::from_listen_address(listen_addr).with_routes(Router::new().fallback_service(service))
 }
 
 async fn build_scrape_response(
